requirements.txt
aiosqlite==0.19.0
anyio==3.6.2
attrs==23.1.0
cattrs==23.1.2
certifi==2022.9.24
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
databases==0.8.0
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.1.3
fastapi==0.87.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
orjson==3.8.2
platformdirs==3.10.0
pydantic==1.10.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.31.0
requests-cache==1.1.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.21.0
typing_extensions==4.4.0
ujson==5.5.0
url-normalize==1.4.3
urllib3==2.0.4
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
